Corrected pom.xml and refactoring errors
