+++ pom.xml
@@ -53,7 +53,7 @@
     <dependency>
       <groupId>com.google.guava</groupId>
       <artifactId>guava</artifactId>
-      <version>19.0-SNAPSHOT</version>
+      <version>19.0</version>
     </dependency>
     <dependency>
       <groupId>junit</groupId>
