+++ src/main/java/io/hops/kafka/ProjectTopics.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package io.hops;
+package io.hops.kafka;
 
 import java.io.Serializable;
 import javax.persistence.Basic;
